BackEnd/requirements.txt
asgiref==3.6.0
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
cryptography==40.0.1
Deprecated==1.2.13
Django==4.1.7
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gcloud==0.18.3
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jwcrypto==1.4.2
Markdown==3.4.3
mypy-extensions==1.0.0
mysqlclient==2.1.1
oauth2client==4.1.3
packaging==23.0
pathspec==0.11.1
platformdirs==3.2.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
PyMySQL==1.0.3
pyparsing==3.0.9
Pyrebase4==4.6.0
python-jwt==4.0.0
pytz==2023.3
requests==2.28.2
requests-toolbelt==0.10.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
